move server code into separate file
