+++ pkg/test/v3/register.go
